Converted to Jekyll
Converted the entire website to use Jekyll, a static-site generator.
This allows me to reuse layouts between HTML pages without copying
code, which makes me happy. I'm also pleased to announce that there
is still no javascript at all in this website, because javascript in
a personal, lightweight website is an abomination.


Former-commit-id: aa74305845d3985f4b7d97e7c5531a73f7e3cffc [formerly aa74305845d3985f4b7d97e7c5531a73f7e3cffc [formerly 6be5f0667a08b66e6e05eaa51540db60b88c6a24 [formerly 2df14f9e12556c872d47f6f968c7068e5c023a26] [formerly 9a97a7de8d8ed2278b047e80e724dd98951c7a56 [formerly 95b317d7d195b9cea1ecf326bf1585da5b63b94c]]]]
Former-commit-id: 85672ae26fa9627d420bf51fbaf39d984b949b7c
Former-commit-id: 941769a048ba4cfbd88e2ff6f1d103d44ed6d8a4
Former-commit-id: 6b41ef0bfe6a4c96b209902457773f969505185c [formerly 7a6b79d533ba7761d2d4db7222deb86933126fe4]
Former-commit-id: bb15c04624cea1c12c23edf7487c3e84cde40d73