Converted to Jekyll
Converted the entire website to use Jekyll, a static-site generator.
This allows me to reuse layouts between HTML pages without copying
code, which makes me happy. I'm also pleased to announce that there
is still no javascript at all in this website, because javascript in
a personal, lightweight website is an abomination.